+++ packages/functions/src/__tests__/falsy.test.ts
@@ -24,7 +24,9 @@ describe('Core Functions / Falsy', () => {
   describe('validation', () => {
     it.each([{}, 2])('given invalid %p options, should throw', async opts => {
       await expect(runFalsy([], opts)).rejects.toThrowAggregateError(
-        new AggregateError([new RulesetValidationError('"falsy" function does not accept any options', [])]),
+        new AggregateError([
+          new RulesetValidationError('invalid-function-options', '"falsy" function does not accept any options', []),
+        ]),
       );
     });
   });