+++ packages/ruleset-migrator/src/validation/schema.ts
@@ -6,44 +6,6 @@ const schema = {
   properties: {
     aliases: {
       type: 'object',
-      additionalProperties: {
-        oneOf: [
-          {
-            type: 'array',
-            items: {
-              type: 'string',
-            },
-          },
-          {
-            type: 'object',
-            properties: {
-              description: {
-                type: 'string',
-              },
-              targets: {
-                type: 'array',
-                minItems: 1,
-                items: {
-                  type: 'object',
-                  properties: {
-                    formats: {
-                      $ref: '#/properties/formats',
-                    },
-                    given: {
-                      type: 'array',
-                      items: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                  required: ['formats', 'given'],
-                },
-              },
-            },
-            required: ['targets'],
-          },
-        ],
-      },
     },
     except: {
       type: 'object',
@@ -85,25 +47,8 @@ const schema = {
     },
     formats: {
       type: 'array',
-      minItems: 1,
       items: {
         type: 'string',
-        enum: [
-          'oas2',
-          'oas3',
-          'oas3.0',
-          'oas3.1',
-          'asyncapi2',
-          'json-schema',
-          'json-schema-loose',
-          'json-schema-draft4',
-          'json-schema-draft6',
-          'json-schema-draft7',
-          'json-schema-draft-2019-09',
-          'json-schema-2019-09',
-          'json-schema-draft-2020-12',
-          'json-schema-2020-12',
-        ],
       },
     },
     functions: {