+++ packages/rulesets/src/asyncapi/functions/asyncApi2MessageExamplesValidation.ts
@@ -1,6 +1,8 @@
 import { createRulesetFunction } from '@stoplight/spectral-core';
 import { schema as schemaFn } from '@stoplight/spectral-functions';
 
+import { mergeTraits } from './utils/mergeTraits';
+
 import type { JsonPath } from '@stoplight/types';
 import type { IFunctionResult, RulesetFunctionContext } from '@stoplight/spectral-core';
 import type { JSONSchema7 } from 'json-schema';
@@ -15,10 +17,11 @@ interface MessageExample {
 export interface MessageFragment {
   payload: unknown;
   headers: unknown;
+  traits?: any[];
   examples?: MessageExample[];
 }
 
-function getMessageExamples(message: MessageFragment): Array<{ path: JsonPath; value: MessageExample }> {
+function getMessageExamples(message: MessageFragment): Array<{ path: JsonPath; example: MessageExample }> {
   if (!Array.isArray(message.examples)) {
     return [];
   }
@@ -26,7 +29,7 @@ function getMessageExamples(message: MessageFragment): Array<{ path: JsonPath; v
     message.examples.map((example, index) => {
       return {
         path: ['examples', index],
-        value: example,
+        example,
       };
     }) ?? []
   );
@@ -68,6 +71,7 @@ export default createRulesetFunction<MessageFragment, null>(
     options: null,
   },
   function asyncApi2MessageExamplesValidation(targetVal, _, ctx) {
+    targetVal = mergeTraits(targetVal); // first merge all traits of message
     if (!targetVal.examples) return;
     const examples = getMessageExamples(targetVal);
 
@@ -75,16 +79,18 @@ export default createRulesetFunction<MessageFragment, null>(
 
     for (const example of examples) {
       // validate payload
-      if (example.value.payload !== undefined) {
-        const errors = validate(example.value.payload, example.path, 'payload', targetVal.payload, ctx);
+      if (example.example.payload !== undefined) {
+        const payload = targetVal.payload ?? {}; // if payload is undefined we treat it as any schema
+        const errors = validate(example.example.payload, example.path, 'payload', payload, ctx);
         if (Array.isArray(errors)) {
           results.push(...errors);
         }
       }
 
       // validate headers
-      if (example.value.headers !== undefined) {
-        const errors = validate(example.value.headers, example.path, 'headers', targetVal.headers, ctx);
+      if (example.example.headers !== undefined) {
+        const headers = targetVal.headers ?? {}; // if headers are undefined we treat them as any schema
+        const errors = validate(example.example.headers, example.path, 'headers', headers, ctx);
         if (Array.isArray(errors)) {
           results.push(...errors);
         }